+++ backend/server.py
@@ -41,6 +41,7 @@ from routers.review_disputes import router as review_disputes_router
 from routers.prod_health import router as prod_health_router
 from routers.updates import router as updates_router
 from routers.coming_soon import router as coming_soon_router
+from routers.growth_stats import router as growth_stats_router
 from seed_data import seed_if_empty
 
 app = FastAPI(title="Crafters Market API")
@@ -82,6 +83,7 @@ api.include_router(review_disputes_router)
 api.include_router(prod_health_router)
 api.include_router(updates_router)
 api.include_router(coming_soon_router)
+api.include_router(growth_stats_router)
 app.include_router(api)
 
 app.add_middleware(